Update cocos2d/base_nodes/CCNode.js
removed bitwise comparison in the context.translate. It caused an issue with scaled layers, because it would convert floats to int, removing the decimal part of a number. This caused animations to run 'chunky', when the container/sprite/node was scaled. I don't know why this was added in the first place, probably to always have a number for the translation.