Fix legacy Function.prototype.bind support
The current legacy support was throwing errors on some mobile devices (specifically an HTC Sense running Android 2.3.4).

The value of args should be determined by the arguments passed to bind, not to the bound function. Several places throughout the codebase bind is used in this manner, so the legacy function should support it.