Fix Firefox audio bug
In the `_play` method, the `offset` parameter can sometimes be `undefined`. This causes the following error in Firefox which effectively disables all audio:

`TypeError: Argument 2 of AudioBufferSourceNode.start is not a finite floating-point value. sourceNode.start(0, offset);`

This pull request fixes the bug by ensuring the offset is 0 if undefined.