Fixed Scale9Sprite setOpacity
Do not propagate opacity to children sprites, in order to avoid multiple opacity levels.
To repro the issue, put on a dark layer a scale9sprite and a stantard sprite loaded with the same texture, the set the opacity to 200. The scale9sprite will be lighter, because of combined opacity of object itself plus children opacity.