Fix SpriteFrame to work with scaled textures
When adding a sprite frame with a scale of <1, the coordinates of the sprite rectangle will (correctly) be larger than the size of the source image. The fix is to check the image size against the pixel dimensions, not the scaled dimensions.

(cherry picked from commit 8f04485a2e938eba0d90ebbd7e3d1adc97cccfe9)